+++ src/main/java/com/programmers/springbasic/controller/VoucherController.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Controller;
 
-import com.programmers.springbasic.controller.dto.CreateFixedAmountVoucherRequest;
-import com.programmers.springbasic.controller.dto.CreatePercentDiscountVoucherRequest;
-import com.programmers.springbasic.controller.dto.ListVouchersResponse;
+import com.programmers.springbasic.dto.CreateFixedAmountVoucherRequest;
+import com.programmers.springbasic.dto.CreatePercentDiscountVoucherRequest;
+import com.programmers.springbasic.dto.ListVouchersResponse;
 import com.programmers.springbasic.service.VoucherService;
 
 @Controller
